* Introducing OD::String, which holds many tools for both FixedString and BufferString * Further reduce the usage of string2.h functions in favor of OD::String or BufferString members * Solved a couple of bugs while at it
git-svn-id: https://svn.opendtect.org/od/trunk@32901 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiTools/uiiosel.cc
@@ -388,7 +388,7 @@ void uiIOSelect::setLabelText( const char* s )
 {
     if ( lbl_ )
     {
-	lbl_->setPrefWidthInChar( FixedString(s).size()+1 );
+	lbl_->setPrefWidthInChar( (int)(FixedString(s).size()+1) );
 	return lbl_->setText( s );
     }
     else if ( optbox_ )
